+++ packages/cli/src/handlers/compile.ts
@@ -2,7 +2,9 @@ import {
     attachTypesToModels,
     convertExplores,
     DbtManifestVersion,
+    getCompiledModels,
     getDbtManifestVersion,
+    getModelsFromManifest,
     getSchemaStructureFromDbtModels,
     isExploreError,
     isSupportedDbtAdapter,
@@ -14,15 +16,10 @@ import { v4 as uuidv4 } from 'uuid';
 import { LightdashAnalytics } from '../analytics/analytics';
 import { getDbtContext } from '../dbt/context';
 import { loadManifest } from '../dbt/manifest';
-import { getModelsFromManifest } from '../dbt/models';
 import { validateDbtModel } from '../dbt/validation';
 import GlobalState from '../globalState';
 import * as styles from '../styles';
-import {
-    DbtCompileOptions,
-    getCompiledModels,
-    maybeCompileModelsAndJoins,
-} from './dbt/compile';
+import { DbtCompileOptions, maybeCompileModelsAndJoins } from './dbt/compile';
 import { getDbtVersion } from './dbt/getDbtVersion';
 import getWarehouseClient from './dbt/getWarehouseClient';
 
@@ -69,7 +66,6 @@ export const compile = async (options: CompileHandlerOptions) => {
             { targetDir: context.targetDir },
             options,
         );
-
     const manifest = await loadManifest({ targetDir: context.targetDir });
     const manifestVersion = getDbtManifestVersion(manifest);
     const manifestModels = getModelsFromManifest(manifest);
