+++ packages/common/src/types/dbt.ts
@@ -571,3 +571,41 @@ export type DbtExposure = {
     url?: string;
     tags?: string[];
 };
+
+export const getModelsFromManifest = (
+    manifest: DbtManifest,
+): DbtModelNode[] => {
+    const models = Object.values(manifest.nodes).filter(
+        (node) =>
+            node.resource_type === 'model' &&
+            node.config?.materialized !== 'ephemeral',
+    ) as DbtRawModelNode[];
+
+    if (!isSupportedDbtAdapter(manifest.metadata)) {
+        throw new ParseError(
+            `dbt adapter not supported. Lightdash does not support adapter ${manifest.metadata.adapter_type}`,
+            {},
+        );
+    }
+    const adapterType = manifest.metadata.adapter_type;
+    return models
+        .filter(
+            (model) =>
+                model.config?.materialized &&
+                model.config.materialized !== 'ephemeral',
+        )
+        .map((model) => normaliseModelDatabase(model, adapterType));
+};
+
+export function getCompiledModels(
+    manifestModels: DbtModelNode[],
+    compiledModelIds?: string[],
+) {
+    return manifestModels.filter((model) => {
+        if (compiledModelIds) {
+            return compiledModelIds.includes(model.unique_id);
+        }
+
+        return model.compiled;
+    });
+}
